Unmarked name field of content gating partition
Unmarked name field of content gating partition. edX mobile apps expect it in english. If value of name field of content gating partition is not `Feature-based Enrollments` string literal mobile apps ignore gating restriction and try to pull content of the block which cause 404 errors for all those blocks which are gated.
ENT-2524