Auto load any keys/values from the yaml config file.
This should allow us to remove all the boilerplate code in this file
where a name is pulled from the config dict and put into the top level
namespace of the settings module.

We do this first so that any logic that adds more complex or dynamic
keys will still run and is safe.

Now that this is here we can start removing any simple boilerplate.