Add configurable Paypal donation button in dahsboard banner
Adds django settings variable PAYPAL_DONATION_HOSTED_BUTTON_ID
which, when set, enables a donation button in the dashboard banner for
recently enrolled audit-track learners to donate to the platform. This
configuration is in addition to the ENABLE_DONATIONS site
configuration and the shoppingcart_donationconfiguration
Config. model, which take precedence. (We plan to remove support for
both of these in the future.)

Implemented as part of the edX Hackathon 23.