Revert "Revert "This stage does the following: (#22692)""
This reverts commit 84de6bc6dec93b561aa9acbb7bbd3165bc3cf5c7
which reverts commit cebeab4348cbe7a288dcfe027a7d448f290651db
which implements the first stage of the `values` column rename in
SiteConfiguration.  However, I included a small change:

This time, we set a default value on the new `site_values` column so
that the ORM will happily deserialize the JSONField without throwing a
JSONDecodeError.