Removing MFE redirect from courseware page.
We decided this approach - of redirecting to the new courseware MFE when users hit the page - won’t be feasible because there are still circumstances when we want that page to load.  For instance, exams, or if an instructor wants to go back to the old experience.  Instead of doing this, we need to track down all the links to the courseware page and put the waffle flag check on those instead.