Add course_visibility to courseware_api.
This data is used in the courseware micro-frontend to determine if the user must be redirected away from a page.  Depending on the visibility value, we may redirect to either the login page, or perhaps the course outline.  Rubric for this is in development, but this piece of data is a pre-requisite.