Fix migration issue with the missing method.
This commit fixes an issue originally in 3541643d where an instance
method on a model was missing in a migration.  The problem was that
Django is smarter than we thought, and is somehow able to construct an
older version of the model before the commit, where there was no such
method.  The solution is just to pull the method out of the model.

DENG-18