Upgrade django-babel-underscore BOM-1069 (#23257)
We upgraded these before in https://github.com/edx/edx-platform/pull/23119 along with django-countries (which released a new version just before that), but had to revert in https://github.com/edx/edx-platform/pull/23237 because of a performance regression in the admin page for EnterpriseCustomer.  We think django-countries was the culprit and have pinned it, so trying these two again by themselves.