Fix docs-generating "make guides" command
The guides documentation generation was broken by two issues:

- A missing CMS-only setting import
- Trying to import the "notes" app from the LMS, which was removed in
Oct. 2019.

I don't know how useful this really is? After all, the command has been
broken for 7 months and no one has complained. Should we simply get rid
of the docs/guides folder? Somehow the docstrings are parsed by sphinx,
and the output is generated as rst files in the cms/lms/openedx/common
folders, but as far as I understand this content is not used anywhere.