Create function to set a Django setting
Currently, Django settings in a production environment are initialized in a
cumbersome and repetitive way, using assignments such as:

VAR_NAME = ENV_TOKENS.get('VAR_NAME', VAR_NAME)

This patch creates a s() function which looks for a variable in the config yaml
and sets it as a global variable in settings. If the variable is not found in the
config yaml, the function uses the global value already assigned to that variable
or an optional default value. This function makes Django settings initialization
shorter and simpler.

This patch also renames additive configurations read from the config yaml such as
FEATURES to EXTRA_FEATURES. Motivation: make clearer to developers and config
yaml editors that these configurations are additive.