Downgrade to pytest-django 3.8.0 for now (#23566)
We'll upgrade this again along with pytest once pytest-dev/pytest#6925 is resolved.

Also, it looks like a flaky error in test class DB setup outside a transaction was mucking things up for the rest of the test shard. I didn't try to fix the flakiness, but this should at least limit the blast radius to just that test class. We really need to start using setUpTestData() more consistently.