Remove the course_id property on student.CourseEnrollment.
It conflicts with an underlying related field on that model which seems
to be getting at the same value from the related table.

Add logging for incorrectly instantiating CourseEnrollment models.
This is to catch any places that might break this that are outside of
edx-platform. Django won't accept `course` values that aren't
course_overviews so we don't need extra logic to test that `course`
values are of the correct type.

fixup! Remove the course_id property on student.CourseEnrollment.