Limit the last_name field to 30 chars.
We don't use this field in edx-platform since we rely on UserProfile
instead.  But we need to do this to prevent accidental incorrect use of
this field.  Django adds a migration between 1.11 and 2.2 to increase
the length of this field that we are opting to skip.  This change
enforces the old limit.