Don't explicitly set the email backend in production.py
We shouldn't be explicitly setting this here. It's already set in all of
our override files and not everyone who wants to use production.py wants
to use SES so we shouldn't hardcode it here.

Also move the other settings to above the config loading so that all
config can be overridden by the config file.