We don't run into this case anymore so remove it.
This was added in here as a shim to catch places that were still using
the old version of these models that had an attribute called course_id
instead of the `course` foreign key.  After monitoring production for
two weeks, we can be pretty confident that we've caught most places that
do this.  Removing this code since it should no longer be needed.