Quiet deprecation warnings from imp module
The "imp" module is deprecated and should be replaced by "importlib". As
a consequence, loading the django settings used to raise deprecation
warnings:

    DeprecationWarning: the imp module is deprecated in favour of
    importlib; see the module's documentation for alternative uses

It should be noted that python 3.5.1 ships with an older release of
distutils which still relies on the imp module. Thus, users of python
3.5.1 (for instance: edx.org developers) will continue to see the
deprecation warning for some time, despite this patch. We suggest
upgrading to python 3.5.9.

This addresses part of CRI-196.