Quiet courseware import warnings
As it is explained in courseware/__init__.py:

    Importing 'lms.djangoapps.courseware' as 'courseware' is no longer
    supported

This warning, while relevant, decreased the signal/noise ratio when
investigating the lms logs.

The "bulk_email" app was the only remaining app importing the courseware
module incorrectly. All edx-platform modules and dependencies now follow
the new convention.

This is in part for CRI-196.