Attempt to migrate the studio video upload view to boto3
We attempt to migrate the "video uploads" view to boto3. This view is
accessible in the CMS via Content -> Video Uploads. It is enabled by
setting the "ENABLE_VIDEO_UPLOAD_PIPELINE" feature flag in the CMS. We
attempted to make this work to make it possible to upload videos to S3
from platforms different than edx.org. However, we realized that we
would need to emulate the entire video processing pipeline.