remove temp_view_func_compare custom metric
In an earlier PR, we moved logic from process_view to process
request, so mapping would happen earlier in the middleware
lifecycle, and the code_owner metric would be set for requests
that never made it to process_view.

The temp_view_func_compare custom metric was added temporarily to
ensure this earlier refactor did not introduce any unaccounted for
differences. It did not, so we are removing this temporary metric.

ARCHBOM-1263