Modify add team count to factor in team visibility
This adds a newly required "user" arg to add_team_count and user context for topic serialization.
Have searched across edx codebase to confirm that nobody else is using this but would be a breaking change for anyone using it externally.

As a note, this increases the number of queries for a team lookup. Hoping to refactor query to do it in one go.