Update topic teams page to query ajax endpoint
The previous code relied on a backbone collection for membership on
teams across the course, while the new logic requires access to teams
only within the teamset.
Was unable to find an easy way to do this with the provided backbone
collections, but a suitable ajax query exists in the API.

In the process of getting this working, I updated much of the local
syntax/formatting to ES6 in an attempt to make the code more
accessible/readable.