reduce some honeycomb tracing on sites middleware
The `CustomDomainsRedirectMiddleware.process_request`,
`RedirectMiddleware.process_request`, `patched_get_site_by_id`, and
`patched_get_site_by_request` functions are now pretty well
understood and optimized. Since those get called on essentially every
request, it adds up to a lot of spans though. This PR removes the
spans (which should help quite a bit on keeping us within our
Honeycomb quota). Some of the fields that were in the spans are still
useful though, so those have been converted to trace fields. That
means they will be available on the top-level trace for each request
and we can easily filter or aggregate on them (eg, get a report of
every request that had a cache miss).