Switch xblock_handler view to  django-rest-framework with Token Auth
The xblock_handler view needs to be able to accept Token
Authentication (OAuth, Jwt) for the upcoming micro-frontend. This PR
makes the change in the view from the plain django-based (using
decorators like `login_required`) and replaces it with the proper
mechanisms from django-rest-framework for authentication and content
negotiation.

The only changes made to the implementation of the view is due to the
fact that django view does not have `request.data` attribute and our
existing request parsing scheme created a `request.json` attribute
with all of the user input. The bulk of the function implementation
and its outputs (possible response objects) remain unchanged, so
there is no new test cases added to this functionality.

https://tasks.opencraft.com/browse/BB-2717