Overriding can_load_courseware if the MFE is disabled for the user
If the user would be allowed to see the courseware MFE (can_load_courseware), we check whether the MFE is disabled for them, based on global settings, course settings (mongo courses), or their particular bucketing in our ExperimentWaffleFlag.

If we determine they shouldn’t be allowed to see it, we return a new CoursewareMicrofrontendDisabledAccessError access response, which the MFE will use to know it should redirect to the old LMS experience.